% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphQC.R
\name{graphQC}
\alias{graphQC}
\title{Representate the compounds area (normalized or not) as a function of their injection order to study trends.}
\usage{
graphQC(LCdata, g, NameDataSet)
}
\arguments{
\item{LCdata}{Matrix of data obtained (mainly by LC-MS) that included four data columns ("Compound Name","Order","QC","Day") and then one coulm for each compound or entity detected (normalized or not).}

\item{g}{Number of compounds for which the graph should be obtained}

\item{NameDataSet}{A name for the data set that is going to be used for the pdf file name. It must be given in quotes}
}
\value{
Multiple graphs of the compounds area (normalized or not) vs the injection order.
}
\description{
Export graphs for each compound included in LCdata matrix in which the area of the specified compound is represented vs the injection order.
}
\examples{
\dontrun{
graphQC(LCdata,3,"datasetName")
}
}

