% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCregresion.R
\name{QCregression}
\alias{QCregression}
\title{Equation to be used internally to predict values from a regression curve of grade 3}
\usage{
QCregression(b, c, d, e, x)
}
\arguments{
\item{b}{coefficient from order 0 part of the equation}

\item{c}{coefficient from order 1 part of the equation}

\item{d}{coefficient from order 2 part of the equation}

\item{e}{coefficient from order 3 part of the equation}

\item{x}{the x-axis value from which the y-axis value wanted to be predicted for the equation given by the coefficients}
}
\value{
A y-value calculated for the x-value especified, taking into account the curve described by the coefficients given
}
\description{
Equation to be used internally to predict values from a regression curve of grade 3
}
\examples{
\dontrun{
prediction<-QCregression(b,c,d,e,x)
}
#' @export
}

