% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCregresion4.R
\name{QCregression4}
\alias{QCregression4}
\title{Equation to be used internally to predict values from a regression curve of grade 4}
\usage{
QCregression4(b, c, d, e, f, x)
}
\arguments{
\item{b}{coefficient from order 0 part of the equation}

\item{c}{coefficient from order 1 part of the equation}

\item{d}{coefficient from order 2 part of the equation}

\item{e}{coefficient from order 3 part of the equation}

\item{f}{coefficient from order 4 part of the equation}

\item{x}{the x-axis value from which the y-axis value wanted to be predicted for the equation given by the coefficients}
}
\value{
A y-value calculated for the x-value especified, taking into account the curve described by the coefficients given
}
\description{
Equation to be used internally to predict values from a regression curve of grade 4
}
\examples{
\dontrun{
prediction<-QCregression4(b,c,d,e,f,x)
}
}

